Add files via upload
As part of a PowerBi Dashboard Dual Challenge in collaboration with TechMeUp and GetResponsive, the challenge was to clean and transform the dataset provided by TechMeUp, which contained real data from all their students past and present. From the data, we had to create a compelling and visually engaging story to present to companies   that could be potential sponsors and employers for their students.